create core dir & separate utils
