+++ src/components/back-panel/index.ts
@@ -1,4 +1,4 @@
-import Block from '../../utils/Block';
+import Block from '../../core/Block';
 import template from './back-panel.hbs';
 
 interface IBackPanelProps {
