+++ src/components/button/index.ts
@@ -1,4 +1,4 @@
-import Block from '../../utils/Block';
+import Block from '../../core/Block';
 import template from './button.hbs';
 
 export interface IButtonProps {
