+++ src/components/chat-item/index.ts
@@ -1,4 +1,4 @@
-import Block from '../../utils/Block';
+import Block from '../../core/Block';
 import template from './chat-item.hbs';
 
 export interface IChatItemProps {
