+++ src/components/input/index.ts
@@ -1,4 +1,4 @@
-import Block from '../../utils/Block';
+import Block from '../../core/Block';
 import template from './input.hbs';
 
 type InputValidationFn = (inputValue: string) => string;
