+++ src/pages/chat/index.ts
@@ -1,4 +1,4 @@
-import Block from '../../utils/Block';
+import Block from '../../core/Block';
 import template from './chat.hbs';
 import { Conversation, IConversation } from '../../components/conversation';
 import { ChatItem, IChatItemProps } from '../../components/chat-item';
