+++ src/pages/login/index.ts
@@ -1,4 +1,4 @@
-import Block from '../../utils/Block';
+import Block from '../../core/Block';
 import template from './login.hbs';
 import { Button } from '../../components/button';
 import { Input, IInputProps } from '../../components/input';
