+++ src/pages/profile/index.ts
@@ -1,4 +1,4 @@
-import Block from '../../utils/Block';
+import Block from '../../core/Block';
 import template from './profile.hbs';
 import { Input, IInputProps } from '../../components/input';
 import { BackPanel } from '../../components/back-panel';
